Conflicts managed and connection corrected ️‍🔥
