+++ lib/Screens/ConfirmationPage/confirmation_screen.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:shelf/Screens/ConfirmationPage/components/body.dart';
 
 class ConfirmationScreen extends StatelessWidget {
-  const ConfirmationScreen({Key? key}) : super(key: key);
+  const ConfirmationScreen({
+    Key? key,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
