+++ lib/Screens/EventsDetailsPage/event_screen.dart
@@ -1,5 +1,5 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:shelf/Screens/EventsPage/components/body.dart';
+import 'package:shelf/Screens/EventsDetailsPage/components/body.dart';
 import 'package:flutter/material.dart';
 import 'package:shelf/Screens/Login/login_screen.dart';
