+++ lib/Screens/RegisteredEventsPage/registered_events.dart
@@ -1,0 +1,11 @@
+import 'package:shelf/Screens/RegisteredEventsPage/components/body.dart';
+import 'package:flutter/material.dart';
+
+class RegisteredEventsPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Body(),
+    );
+  }
+}
